feat: add delete and update end point

+++ src/orga_structure/application/command/handlers/DeleteRoleHandler.ts
@@ -1,8 +1,8 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { DeleteRoleCommand } from '../DeleteRoleCommand';
 import { RolesRepository } from 'src/orga_structure/domain/role/RoleRepository';
-import { InjectionTokens } from '../../InjectionTokens';
-import { Inject } from '@nestjs/common';
+import { InjectionTokens } from 'src/shared/InjectionTokens';
+import { Inject, NotFoundException } from '@nestjs/common';
 
 @CommandHandler(DeleteRoleCommand)
 export class DeleteRoleHandler
@@ -14,6 +14,11 @@ export class DeleteRoleHandler
   ) {}
 
   async execute(command: DeleteRoleCommand): Promise<void> {
-    await this.rolesRepository.delete(command.id);
+    const role = await this.rolesRepository.findById(command.id);
+    if (role) {
+      await this.rolesRepository.delete(command.id);
+    } else {
+      throw new NotFoundException('Role not found.');
+    }
   }
 }
